+++ src/shell_can.h
@@ -1,7 +1,6 @@
 #ifndef _SHELL_CAN_H
 #define _SHELL_CAN_H
 
-
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -33,7 +32,7 @@ struct can_command
 
 };
 
-int8_t can_shell_handler(char *cmd, uint8_t len);
+int can_shell_handler(char *cmd, uint8_t len);
 
 
 #ifdef __cplusplus
